First and last payments

+++ packages/rfv/interactions/membership.ts
@@ -8,7 +8,7 @@ const apihelper = new JXPHelper({ server: config.api.server, apikey: process.env
 export async function get_memberships(mday: moment.Moment) {
     try {
         console.log("get_memberships", mday.format("YYYY-MM-DD"));
-        const interactions = await mongo.find("interactions", { day: mday.format("YYYY-MM-DD") });
+        const interactions = await mongo.find("interactions", { day: mday.toDate() });
         // console.log(interactions.slice(0, 10));
         const reader_ids = interactions.map(i => i.reader_id);
         // console.log(reader_ids.slice(0, 10));
@@ -32,23 +32,20 @@ export async function get_memberships(mday: moment.Moment) {
         // console.log(JSON.stringify(reader_pipeline, null, 2));
         const readers = await mongo.aggregate("readers", reader_pipeline);
         console.timeEnd("get readers");
-        console.log("Getting orders...")
+        console.log(`Readers: ${readers.length}`);
         console.time("get orders");
         const customer_ids = readers.map(r => r.wordpress_id);
+        // Get previous month's orders
+        const previous_month = mday.clone().subtract(1, "month");
         const orders_pipeline = [
             {
                 $match: {
                     customer_id: { $in: customer_ids },
                     date_paid: {
-                        $lte: mday.clone().endOf("month").toDate(),
+                        $lte: previous_month.clone().endOf("month").toDate(),
                     }
                 }
             },
-            // {
-            //     $sort: {
-            //         date_paid: -1
-            //     }
-            // },
             {
                 $group: {
                     _id: "$customer_id",
@@ -61,18 +58,63 @@ export async function get_memberships(mday: moment.Moment) {
                 $match: {
                     dates_paid: {
                         $ne: null,
-                        $lte: new Date(mday.clone().endOf("month").toDate()),
-                        $gte: new Date(mday.clone().startOf("month").toDate())
+                        $lte: new Date(previous_month.clone().endOf("month").toDate()),
+                        $gte: new Date(previous_month.clone().startOf("month").toDate())
                     },
                 }
+            },
+            // First and last payments
+            {
+                $lookup: {
+                    from: "woocommerce_orders",
+                    let: { customer_id: "$_id" },
+                    pipeline: [
+                        {
+                            $match: {
+                                $expr: {
+                                    $eq: ["$customer_id", "$$customer_id"]
+                                },
+                                date_paid: {
+                                    $ne: null,
+                                    $exists: true,
+                                    $lte: previous_month.clone().endOf("month").toDate(),
+                                }
+                            }
+                        },
+                        {
+                            $sort: {
+                                date_paid: 1
+                            }
+                        },
+                        {
+                            $group: {
+                                _id: "$customer_id",
+                                first_payment: { $first: "$date_paid" },
+                                last_payment: { $last: "$date_paid" }
+                            }
+                        },
+                        {
+                            $project: {
+                                _id: 0,
+                                wordpress_id: "$_id",
+                                first_payment: 1,
+                                last_payment: 1
+                            }
+                        }
+                    ],
+                    as: "first_last_payments"
+                }
+            },
+            {
+                $unwind: "$first_last_payments"
             },
             {
                 $project: {
                     _id: 0,
-                    customer_id: "$_id",
+                    wordpress_id: "$_id",
                     lifetime_value: 1,
                     monthly_value: "$total",
-                    insider: true,
+                    insider: { $literal: true },
                     date_paid: { 
                         $first: {
                             $filter: { 
@@ -80,13 +122,15 @@ export async function get_memberships(mday: moment.Moment) {
                                 as: "date_paid",
                                 cond: { 
                                     $and: [
-                                        { $lte: ["$$date_paid", mday.clone().endOf("month").toDate()] },
-                                        { $gte: ["$$date_paid", mday.clone().startOf("month").toDate()] }
+                                        { $lte: ["$$date_paid", previous_month.clone().endOf("month").toDate()] },
+                                        { $gte: ["$$date_paid", previous_month.clone().startOf("month").toDate()] }
                                     ]
                                 } 
                             }
                         }
-                    }
+                    },
+                    first_payment: "$first_last_payments.first_payment",
+                    last_payment: "$first_last_payments.last_payment"
                 }
             }
         ];
@@ -94,12 +138,17 @@ export async function get_memberships(mday: moment.Moment) {
         const orders = await mongo.aggregate("woocommerce_orders", orders_pipeline);
         // console.log(orders.slice(0, 10));
         console.timeEnd("get orders");
+        console.log(`Orders: ${orders.length}`);
         for (let order of orders) {
-            const email = readers.find(r => r.wordpress_id === order.customer_id).email;
+            const email = readers.find(r => r.wordpress_id === order.wordpress_id)?.email;
+            if (!email) {
+                console.log("No email for", order.wordpress_id);
+                continue;
+            }
             order.uid = `${email}-${mday.format("YYYY-MM-DD")}`;
-            delete order.customer_id;
+            // delete order.wordpress_id;
         }
-        console.log(orders.slice(0, 10));
+        // console.log(orders.slice(0, 10));
         return orders;
     } catch (err) {
         console.error(err);
@@ -110,6 +159,10 @@ export async function get_memberships(mday: moment.Moment) {
 
 export async function process_memberships(mday: moment.Moment) {
     const interactions = await get_memberships(mday);
+    if (interactions.length === 0) {
+        console.log("No interactions found for", mday.format("YYYY-MM-DD"));
+        return;
+    }
     const tmp_collection = `tmp_memberships_${mday.format("YYYYMMDD")}`;
     if (await mongo.collectionExists(tmp_collection)) {
         await mongo.dropCollection(tmp_collection);
@@ -124,7 +177,8 @@ export async function process_memberships(mday: moment.Moment) {
                 lifetime_value: 1,
                 monthly_value: 1,
                 date_paid: 1,
-                insider: 1
+                insider: 1,
+                wordpress_id: 1
             }
         },
         {
@@ -137,5 +191,40 @@ export async function process_memberships(mday: moment.Moment) {
         }
     ];
     await mongo.aggregate(tmp_collection, merge_query);
+    await mongo.dropCollection(tmp_collection);
     await mongo.close();
+}
+
+export function first_last_payments() {
+    try {
+        const missing_dates_query = mongo.find("interactions", { date_paid: { $exists: true },  first_payment: { $exists: false }, last_payment: { $exists: false } }, { wordpress_id: true });
+        const ids = missing_dates_query.map(i => i.wordpress_id);
+        const unique_ids = new Set(ids);
+        const pipeline = [
+            {
+                $match: {
+                    customer_id: { $in: unique_ids },
+                    date_paid: { $exists: true, $ne: null }
+                }
+            },
+            {
+                $group: {
+                    _id: "$customer_id",
+                    first_payment: { $min: "$date_paid" },
+                    last_payment: { $max: "$date_paid" }
+                }
+            },
+            {
+                $project: {
+                    _id: 0,
+                    wordpress_id: "$_id",
+                    first_payment: 1,
+                    last_payment: 1
+                }
+            }
+        ];
+        const first_last_payments = mongo.aggregate("woocommerce_orders", pipeline);
+    } catch (err) {
+        console.error(err);
+    }
 }
